requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyserial>=3.5
pyyaml>=6.0
flask>=2.3.0
